Simplify a little
